+++ R/super.R
@@ -37,7 +37,7 @@
 #' method(total, foo2) <- function(x) total(super(x, foo1)) + x@z
 #' total(foo2(1, 2, 3))
 #'
-#' # To see the difference between cast() and super() we need a
+#' # To see the difference between convert() and super() we need a
 #' # method that calls another generic
 #'
 #' bar1 <- new_generic("bar1", "x")
@@ -50,8 +50,8 @@
 #'
 #' obj <- foo2(1, 2, 3)
 #' bar2(obj)
-#' # cast() affects every generic:
-#' bar2(cast(obj, foo1))
+#' # convert() affects every generic:
+#' bar2(convert(obj, foo1))
 #' # super() only affects the _next_ generic:
 #' bar2(super(obj, foo1))
 super <- function(from, to) {
@@ -61,7 +61,7 @@ super <- function(from, to) {
   check_can_inherit(to)
   if (!class_inherits(from, to)) {
     msg <- sprintf(
-      "Can't cast: %s doesn't inherit from %s",
+      "%s doesn't inherit from %s",
       obj_desc(from),
       class_desc(to)
     )