+++ tests/testthat/test-super.R
@@ -26,7 +26,7 @@ describe("super()", {
     method(bar2, foo2) <- function(x) c(2, bar1(x))
 
     expect_equal(bar2(super(foo2(), foo1)), c(1, 2))
-    expect_equal(bar2(cast(foo2(), foo1)), c(1, 1))
+    expect_equal(bar2(convert(foo2(), foo1)), c(1, 1))
   })
 
   it("checks to", {